+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add `FlowbiteBreadcrumb` component.
 - Add `FlowbiteToast` component.
+- Add `FlowbiteProgressBar` component.
 
 ## [0.0.5] - 11-09-2026
 ### Added